feat: cash balance services impl + left OpenApi modules

+++ TradeBot/src/main/java/com/domingueti/tradebot/modules/CashBalance/services/DeleteCashBalanceByIdService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.domingueti.tradebot.exceptions.NotFoundException;
+import com.domingueti.tradebot.modules.CashBalance.models.CashBalance;
 import com.domingueti.tradebot.modules.CashBalance.repositories.CashBalanceRepository;
 
 @Service
@@ -14,8 +16,17 @@ public class DeleteCashBalanceByIdService {
 	
 	@Transactional
 	public void execute(Long id) {
-//		validator.execute(id); 
+		
+		CashBalance cashBalance = cashBalanceRepository.findByIdAndDeletedAtIsNull(id);
+		
+//		validator.execute(cashBalance);
+		
+		if (cashBalance == null) {
+			throw new NotFoundException("Cash balance not found with given ID: " + id);
+		}
+
 		cashBalanceRepository.deleteById(id);
+	
 	}
 	
 }
